Refactored PomRewriter
